+++ src/Model/Snapshot.h
@@ -12,11 +12,11 @@ public:
     virtual ~Snapshot() = default;
 
     const char *getName(void) const;
-    uint8_t getColour(void) const;
+    uint16_t getColour(void) const;
 
 private:
     bool parse(File &file);
 
     char name[MAX_SNAPSHOT_NAME_LENGTH + 1];
-    uint8_t colour;
+    uint16_t colour;
 };
